console_app: use custom device builder to init devices
Each time the USB library wants to create a new Device,
the device_creator_func is called. This allows the application to
override the Device implementation (e.g. return a subclassed version).

Apart from the possibility to subclass Device, the device may also be
configured at this time, for example to subscribe to certain events etc.

From now on, this is the recommended way to initialize a Device.