Add CallbackQueue: synchronize callbacks to GUI thread
A CallbackQueue instance can wrap any function. Each time the wrapped
function is called, the original function is not actually called
directly, but submitted to an internal queue.

When the queue is polled, the original function(s) are called (with the
original paramater(s)).

Note: the wrapped functions always return None, as the function is
effectively asynchronous and the returnvalue is not known yet.